Add doc input `Search::class`.
