+++ tests/Input/Search/AttributeTest.php
@@ -167,26 +167,6 @@ final class AttributeTest extends TestCase
         );
     }
 
-    public function testMaxLength(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="search-6582f2d099e8b" type="search" maxlength="1">
-            HTML,
-            Search::widget()->id('search-6582f2d099e8b')->maxlength(1)->render()
-        );
-    }
-
-    public function testMinLength(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="search-6582f2d099e8b" type="search" minlength="1">
-            HTML,
-            Search::widget()->id('search-6582f2d099e8b')->minlength(1)->render()
-        );
-    }
-
     public function testName(): void
     {
         Assert::equalsWithoutLE(
@@ -194,16 +174,6 @@ final class AttributeTest extends TestCase
             <input id="search-6582f2d099e8b" name="value" type="search">
             HTML,
             Search::widget()->id('search-6582f2d099e8b')->name('value')->render()
-        );
-    }
-
-    public function testPattern(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="search-6582f2d099e8b" type="search" pattern="value">
-            HTML,
-            Search::widget()->id('search-6582f2d099e8b')->pattern('value')->render()
         );
     }
 
@@ -224,26 +194,6 @@ final class AttributeTest extends TestCase
             <input id="search-6582f2d099e8b" type="search" readonly>
             HTML,
             Search::widget()->id('search-6582f2d099e8b')->readonly()->render()
-        );
-    }
-
-    public function testRender(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="search-6582f2d099e8b" type="search">
-            HTML,
-            Search::widget()->id('search-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testRequired(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="search-6582f2d099e8b" type="search" required>
-            HTML,
-            Search::widget()->id('search-6582f2d099e8b')->required()->render()
         );
     }
